% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificar_bairros_distrito_novo.R
\name{atribuir_distrito_proximo}
\alias{atribuir_distrito_proximo}
\title{Atribuir distrito mais próximo para bairros sem interseção}
\usage{
atribuir_distrito_proximo(sp_bairros, sp_distritos)
}
\arguments{
\item{sp_bairros}{Objeto \code{sf} contendo bairros.}

\item{sp_distritos}{Objeto \code{sf} contendo distritos.}
}
\value{
Objeto \code{sf} com coluna \code{distrito} preenchida.
}
\description{
Para bairros que não foram associados a um distrito por interseção
espacial, atribui o distrito mais próximo utilizando
\code{sf::st_nearest_feature()}.
}
\examples{
\dontrun{
bairros <- atribuir_distrito_proximo(bairros, sp_distritos)
}

}
