% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_alerta_novo.R
\name{calcular_alerta_cidade}
\alias{calcular_alerta_cidade}
\title{Calcular alertas epidemiológicos para todas as unidades de uma cidade}
\usage{
calcular_alerta_cidade(
  data,
  coluna_id,
  params,
  gtdist = "normal",
  meangt = 3,
  sdgt = 1,
  criteria_fn = setCriteria
)
}
\arguments{
\item{data}{Data frame contendo os dados epidemiológicos.}

\item{coluna_id}{Nome da coluna de identificação geográfica.}

\item{params}{Lista de parâmetros do modelo de alerta.}

\item{gtdist}{Distribuição do tempo de geração.}

\item{meangt}{Média do tempo de geração.}

\item{sdgt}{Desvio padrão do tempo de geração.}
}
\value{
Data frame contendo os níveis de alerta para todas as unidades
geográficas.
}
\description{
Executa o cálculo de alerta epidemiológico para todas as unidades
geográficas presentes em um dataset (por exemplo, bairros ou distritos).
}
\examples{
\dontrun{
calcular_alerta_cidade(
  data = dados_bairros,
  coluna_id = "id_bairro",
  params = params_modelo
)
}

}
