% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificar_bairros_distrito_novo.R
\name{calcular_centroides}
\alias{calcular_centroides}
\title{Calcular centróides de polígonos}
\usage{
calcular_centroides(sp_obj)
}
\arguments{
\item{sp_obj}{Objeto \code{sf} contendo geometria poligonal.}
}
\value{
Objeto \code{sf} contendo os centróides.
}
\description{
Calcula os centróides de um objeto espacial \code{sf}.
}
\examples{
\dontrun{
centroides <- calcular_centroides(sp_bairros)
}

}
