% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_alerta_novo.R
\name{calcular_rt}
\alias{calcular_rt}
\title{Calcular número reprodutivo efetivo (Rt)}
\usage{
calcular_rt(data, gtdist = "normal", meangt = 3, sdgt = 1)
}
\arguments{
\item{data}{Data frame contendo os dados epidemiológicos.}

\item{gtdist}{Distribuição do tempo de geração.}

\item{meangt}{Média do tempo de geração.}

\item{sdgt}{Desvio padrão do tempo de geração.}
}
\value{
Data frame com estimativas de Rt adicionadas.
}
\description{
Aplica o modelo de estimativa do número reprodutivo efetivo (Rt)
utilizando a função \code{AlertTools::Rt()}.
}
\examples{
\dontrun{
calcular_rt(dados_bairro)
}

}
