% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criar_breaks_1ano.R
\name{criar_breaks_1ano}
\alias{criar_breaks_1ano}
\title{Criar breaks do eixo X para janela móvel de 1 ano}
\usage{
criar_breaks_1ano(ano_fim, semana_fim, passo = 10)
}
\arguments{
\item{ano_fim}{Inteiro. Ano da última semana da série.}

\item{semana_fim}{Inteiro. Semana epidemiológica final (1–52).}

\item{passo}{Inteiro. Intervalo entre os breaks em semanas.
Default é \code{10}.}
}
\value{
Um vetor numérico contendo os breaks do eixo X no formato \code{YYYYWW}.
}
\description{
Gera pontos de quebra (breaks) para o eixo X de gráficos baseados em
semanas epidemiológicas, considerando uma janela móvel de 1 ano
(52 semanas). Os breaks são calculados em intervalos definidos
pelo argumento \code{passo} e retornados no formato \code{YYYYWW}.
}
\details{
A função converte o ano e a semana final em um índice absoluto de
semanas, calcula os pontos de quebra dentro da janela de 52 semanas
anteriores e depois reconverte para o formato \code{ano-semana}.
}
\examples{
\dontrun{
# Gerar breaks para uma série que termina na semana 35 de 2024
criar_breaks_1ano(
  ano_fim = 2024,
  semana_fim = 35
)

# Usando passo menor
criar_breaks_1ano(
  ano_fim = 2024,
  semana_fim = 35,
  passo = 5
)
}

}
