% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcular_alerta_novo.R
\name{filtrar_unidade}
\alias{filtrar_unidade}
\title{Filtrar dados de uma unidade geográfica}
\usage{
filtrar_unidade(data, coluna_id, id)
}
\arguments{
\item{data}{Data frame contendo os dados epidemiológicos.}

\item{coluna_id}{String com o nome da coluna que identifica a unidade
geográfica.}

\item{id}{Valor do identificador da unidade geográfica.}
}
\value{
Data frame filtrado para a unidade geográfica e ordenado por
\code{sem_not}.
}
\description{
Filtra o conjunto de dados epidemiológicos para uma unidade geográfica
específica e ordena as observações pela semana epidemiológica.
}
\examples{
\dontrun{
filtrar_unidade(dados, "id_bairro", 101)
}

}
