% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infodengesubmunicipal-package.R
\docType{package}
\name{infodengesubmunicipal-package}
\alias{infodengesubmunicipal}
\alias{infodengesubmunicipal-package}
\title{infodengesubmunicipal: ferramentas para analise epidemiologica submunicipal}
\description{
Infraestrutura de namespace do pacote. Os imports centralizados aqui
refletem o uso atual do codigo e reduzem erros de carregamento enquanto
a refatoracao das funcoes segue em andamento.
}
\author{
\strong{Maintainer}: Quemuel Baruque \email{quemuelbaruque@gmail.com}

}
\keyword{internal}
