% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificar_bairros_distrito_novo.R
\name{intersectar_distritos}
\alias{intersectar_distritos}
\title{Intersectar bairros e distritos via centróides}
\usage{
intersectar_distritos(centroides_bairros, sp_distritos)
}
\arguments{
\item{centroides_bairros}{Objeto \code{sf} contendo centróides dos bairros.}

\item{sp_distritos}{Objeto \code{sf} contendo polígonos de distritos.}
}
\value{
Data frame com colunas \code{nome_bairr}, \code{id_bairro} e \code{distrito}.
}
\description{
Identifica o distrito correspondente a cada bairro utilizando
a interseção espacial entre os centróides dos bairros e os
polígonos de distritos.
}
\examples{
\dontrun{
intersecao <- intersectar_distritos(centroides, sp_distritos)
}

}
