% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_dados_dengue.R
\name{ler_dados_dengue}
\alias{ler_dados_dengue}
\title{Lê e filtra dados históricos de dengue}
\usage{
ler_dados_dengue(caminho)
}
\arguments{
\item{caminho}{Caminho para o arquivo DBF contendo os dados.}
}
\value{
Um \code{data.frame} contendo as colunas:
\itemize{
\item NU_NOTIFIC
\item NU_ANO
\item SEM_NOT
\item SG_UF_NOT
\item ID_UNIDADE
\item ID_BAIRRO
\item NM_BAIRRO
\item ID_MN_RESI
\item CLASSI_FIN
\item DT_SIN_PRI
\item DT_DIGITA
}
}
\description{
Lê arquivo DBF contendo notificações de dengue
e retorna subconjunto filtrado por município
e com colunas selecionadas.
}
\details{
Atualmente o filtro do município está fixado para o código
IBGE "420910". Recomenda-se refatorar a função para
receber o código como parâmetro para permitir escalabilidade.
}
\examples{
\dontrun{
dados <- ler_dados_dengue("dados_dengue.dbf")
}

}
